Utils: documentation generator has now a base object
